+++ packages/antd/src/components/pages/auth/components/register/index.tsx
@@ -158,7 +158,7 @@ export const RegisterPage: React.FC<RegisterProps> = ({
                         >
                             {translate(
                                 "pages.login.buttons.haveAccount",
-                                "Have an account? ",
+                                "Have an account?",
                             )}{" "}
                             <Link
                                 style={{