+++ packages/mui/src/components/pages/auth/components/styles.ts
@@ -1,0 +1,18 @@
+import { CSSProperties } from "react";
+
+export const layoutStyles: CSSProperties = {
+    background: `radial-gradient(50% 50% at 50% 50%, #6813CB 0%, #2A2A42 100%)`,
+    backgroundSize: "cover",
+};
+
+export const titleStyles: CSSProperties = {
+    textAlign: "center",
+    color: "#626262",
+    fontSize: "24px",
+    letterSpacing: "-0.04em",
+    marginBottom: "24px",
+    overflowWrap: "break-word",
+    hyphens: "manual",
+    textOverflow: "unset",
+    whiteSpace: "pre-wrap",
+};